requirements-app.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
stripe
python-dotenv
Authlib
requests
matplotlib
pandas
pytest
redis
